# Conflicts:
#	src/client/Presentation/AccountantUI/AccountManagementUI.java

+++ src/client/BL/Accountant/FinancialCheckSellbl/FinancialCheckSellController.java
@@ -45,9 +45,9 @@ public class FinancialCheckSellController implements FinancialCheckSellInterface
         String saleTime = vo.getDateandtime();
         String goodsName =goodslist.getGoodsname();
         String goodsType =goodslist.getKeymodel();
-        Long goodsNum = goodslist.getNum();
-        Long goodsPrice = goodslist.getPrice();
-        Long totalSum = goodsNum*goodsPrice;
+        double goodsNum = goodslist.getNum();
+        double goodsPrice = goodslist.getPrice();
+        double totalSum = goodsNum*goodsPrice;
 
         return null;
     }
